+++ include/inc_lang/date/pl.date.lang.php
@@ -1,50 +1,37 @@
 <?php
-/*************************************************************************************
-   Copyright notice
-   
-   (c) 2004 Oliver Georgi (oliver@xletter.de, http://www.xletter.org)
-   All rights reserved.
- 
-   This script is part of xLetter - the PHP newsletter & mailing list software.
-   
-   xLetter is free software; you can redistribute it and/or modify it under the terms
-   of the GNU General Public License as published by the Free Software Foundation;
-   either version 2 of the License, or (at your option) any later version.
-  
-   The GNU General Public License can be found at http://www.gnu.org/copyleft/gpl.html
-   A copy is found in the textfile GPL.txt and important notices to the license 
-   from the author is found in LICENSE.txt distributed with these scripts.
-  
-   This script is distributed in the hope that it will be useful, but WITHOUT ANY 
-   WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
-   PARTICULAR PURPOSE.  See the GNU General Public License for more details.
- 
-   This copyright notice MUST APPEAR in all copies of the script!
-*************************************************************************************/
+/**
+ * phpwcms content management system
+ *
+ * @author Oliver Georgi <oliver@phpwcms.de>
+ * @copyright Copyright (c) 2002-2012, Oliver Georgi
+ * @license http://opensource.org/licenses/GPL-2.0 GNU GPL-2
+ * @link http://www.phpwcms.de
+ *
+ **/
 
 // pl.date.lang.php
-// rev 1.0; 01-14-2004
+// Language: Polish, Language Code: pl UTF-8 for ver.1.6.515 Zibi http://www.krynica.malopolska.pl
 
 // language: PL
 
-$weekday_long[0]	= "Niedziela";		$weekday_short[0]	= "Nd";
-$weekday_long[1]	= "Poniedzia³ek";	$weekday_short[1]	= "Pn";
-$weekday_long[2]	= "Wtorek";		$weekday_short[2]	= "Wt";
-$weekday_long[3]	= "¦roda";		$weekday_short[3]	= "¦r";
-$weekday_long[4]	= "Czwartek";		$weekday_short[4]	= "Cz";
-$weekday_long[5]	= "Pi±tek";		$weekday_short[5]	= "Pi";
-$weekday_long[6]	= "Sobota";		$weekday_short[6]	= "So";
-$month_long[ 1]	= "Styczeñ";			$month_short[ 1]	= "Sty.";
-$month_long[ 2]	= "Luty";		        $month_short[ 2]	= "Lut.";
+$weekday_long[0] = "Niedziela";		$weekday_short[0]	= "Nd";
+$weekday_long[1] = "PoniedziaÅ‚ek";	$weekday_short[1]	= "Pn";
+$weekday_long[2] = "Wtorek";		$weekday_short[2]	= "Wt";
+$weekday_long[3] = "Åšroda";			$weekday_short[3]	= "Åšr";
+$weekday_long[4] = "Czwartek";		$weekday_short[4]	= "Cz";
+$weekday_long[5] = "PiÄ…tek";		$weekday_short[5]	= "Pi";
+$weekday_long[6] = "Sobota";		$weekday_short[6]	= "So";
+$month_long[ 1]	= "StyczeÅ„";		$month_short[ 1]	= "Sty.";
+$month_long[ 2]	= "Luty";			$month_short[ 2]	= "Lut.";
 $month_long[ 3]	= "Marzec";			$month_short[ 3]	= "Marz";
-$month_long[ 4]	= "Kwiecieñ";			$month_short[ 4]	= "Kwi.";
+$month_long[ 4]	= "KwiecieÅ„";		$month_short[ 4]	= "Kwi.";
 $month_long[ 5]	= "Maj";			$month_short[ 5]	= "Maj";
-$month_long[ 6]	= "Czerwiec";			$month_short[12]	= "Cze.";
+$month_long[ 6]	= "Czerwiec";		$month_short[12]	= "Cze.";
 $month_long[ 7]	= "Lipiec";			$month_short[ 6]	= "Lip.";
-$month_long[ 8]	= "Sierpieñ";			$month_short[ 7]	= "Sier.";
-$month_long[ 9]	= "Wrzesieñ";			$month_short[ 8]	= "Wrz.";
-$month_long[10]	= "Pa¼dziernik";		$month_short[ 9]	= "Pa¼.";
-$month_long[11]	= "Listopad";			$month_short[10]	= "Lis.";
-$month_long[12]	= "Grudzieñ";			$month_short[11]	= "Gru.";
+$month_long[ 8]	= "SierpieÅ„";		$month_short[ 7]	= "Sier.";
+$month_long[ 9]	= "WrzesieÅ„";		$month_short[ 8]	= "Wrz.";
+$month_long[10]	= "PaÅºdziernik";	$month_short[ 9]	= "PaÅº.";
+$month_long[11]	= "Listopad";		$month_short[10]	= "Lis.";
+$month_long[12]	= "GrudzieÅ„";		$month_short[11]	= "Gru.";
 
 ?>